requirements.txt
absl-py==2.1.0
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
bidict==0.22.1
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.0
decorator==5.1.1
decord==0.6.0
dlib==19.24.2
easydict==1.11
executing==2.0.1
filelock==3.9.0
Flask==3.0.2
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2023.12.2
gast==0.4.0
gevent==23.9.1
google-auth==2.27.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.4
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
imutils==0.5.4
insightface==0.7.3
ipykernel==6.29.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lapx==0.5.5
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
-e git+https://github.com/WildChlamydia/MiVOLO.git@57581b7da623d19081b341ea1a99798ca87b728f#egg=mivolo
mpmath==1.3.0
mtcnn==0.1.1
mutagen==1.47.0
nest-asyncio==1.6.0
networkx==3.0
numba==0.58.1
numpy==1.24.1
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.17.0
onnxruntime-gpu==1.17.0
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pooch==1.8.0
prettytable==3.9.0
prompt-toolkit==3.0.43
protobuf==3.20.2
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.20.0
Pygments==2.17.2
PyMatting==1.1.12
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qudida==0.0.4
referencing==0.33.0
rembg==2.0.54
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
-e git+https://github.com/andresprados/SPIGA.git@a709c95cc93f8246d7bff4cfb970a2dd0d62ffed#egg=spiga
stack-data==0.6.3
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.8.13.dev0
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.8.0
tzdata==2023.4
ultralytics==8.0.124
urllib3==2.2.0
waitress==3.0.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yt-dlp==2023.12.30
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
